Fourier coefficients

Fast Fourier Transform (FFT) of a series of temperatures.

Usage

fft_rfft(s)

Arguments

s

num. Sampled temperatures at constant interval.

Value

\(f\) the Fourier coefficients of \(s\) of length \(\ell\), whose coefficients are \(c_{0},c_{1},\ldots,c_{\ell-1}\)

Details

A series \(S\) consisting of \(N\) samples measured at a time interval \(\Delta t\) can be analyzed by a Fourier series

$$s(t) = a_{0} + \sum_{n=1}^{\ell} a_{n} \cos(2\pi \tfrac{n}{T} t) + b_{n}\sin(2\pi \tfrac n T t)$$

where \(a_n\) and \(b_n\) are the Fourier coefficients obtained from the fast fourier transform (FFT) of \(S\). Note, the time reference of the first measure is \(t=0\).

The number of coeffcients is $$ \ell = \begin{cases} N/2 + 1 & \text{if } N \text{ is even} \\ (N+1)/2 & \text{if } N \text{ is odd} \end{cases} $$

Generally they take the form of complex number $$c_{n} = a_{n} - i b_{n} \,, \quad n=0,\ldots, \ell-1$$ with the convention \(b_{0}=0\). Hence, we have the relation $$a_{n} = \Re (c_{n}) \text{ and } b_{n} = - \Im(c_{n})\,, \quad n=0,\ldots,\ell \,.$$

Examples

fft_rfft(hobo$t_hobo[1:(24*5)])
#>  [1] 17.619925000+0.000000000i  0.352106589+1.112782143i
#>  [3]  0.016938427+0.305651468i  0.786822892-0.491193725i
#>  [5] -0.366091153+0.405903371i  0.561862627-0.705511083i
#>  [7] -0.700440977+0.909059469i -0.129852090+0.151315838i
#>  [9]  0.259901949+0.262260371i -0.835430549+0.246510019i
#> [11]  0.195387351-0.651666043i  0.328159272+0.068455153i
#> [13] -0.110472830+0.715487848i  0.747517003-0.561106310i
#> [15]  0.024473573+0.122845179i -0.171405238-0.398025340i
#> [17]  0.507700502+0.369399474i  0.503616617+0.118809570i
#> [19] -0.386323061-0.245562345i  0.100025132-0.277725369i
#> [21]  0.344575000+0.478839880i  0.054695254-0.260030182i
#> [23] -0.010326392-0.050341598i -0.081273749-0.004086432i
#> [25]  0.159666620-0.149883773i  0.025848915+0.267768576i
#> [27]  0.518948895-0.282354686i -0.117936182+0.172249986i
#> [29] -0.017146601-0.557189207i -0.002634857+0.216013547i
#> [31] -0.123350000-0.397733333i  0.569067855+0.089021465i
#> [33] -0.248775354-0.081325972i -0.235441227-0.524633870i
#> [35] -0.144549093-0.122659407i -0.164967039-0.060847625i
#> [37] -0.029527170+0.425698636i -0.170526243-0.108091891i
#> [39]  0.155599885-0.208123915i -0.018751128-0.400831657i
#> [41] -0.421025000+0.367902025i -0.286435814+0.266492324i
#> [43]  0.222948397-0.152273358i -0.063420836-0.034134178i
#> [45]  0.024617441+0.374621548i  0.033938571-0.073992007i
#> [47] -0.073998916-0.240159504i -0.215262917+0.424945127i
#> [49]  0.094783380+0.110737052i  0.277095763-0.075886821i
#> [51] -0.269437351-0.073767291i  0.153397293+0.360540865i
#> [53]  0.025170313+0.075429836i -0.278453672-0.242105323i
#> [55]  0.319832308+0.418421495i  0.207422164+0.338638451i
#> [57] -0.211177097-0.276677170i  0.047443647+0.283219503i
#> [59]  0.610546955+0.036349453i  0.164771944-0.168584366i
#> [61] -0.554750000+0.000000000i